pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ai-code-reviewer"
version = "0.1.0"
description = "AI Code Reviewer using CodeBERT for bug detection"
authors = [{name = "AI Code Reviewer Team"}]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "torch>=2.1.0",
    "transformers>=4.35.0",
    "datasets>=2.14.5",
    "evaluate>=0.4.1",
    "accelerate>=0.24.1",
    "pandas>=2.1.3",
    "pyarrow>=14.0.1",
    "numpy>=1.24.3",
    "scikit-learn>=1.3.2",
    "tqdm>=4.66.1",
    "fastapi>=0.104.1",
    "uvicorn>=0.24.0",
    "pydantic>=2.5.0",
    "python-multipart>=0.0.6",
    "streamlit>=1.28.1",
    "pyyaml>=6.0.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.3",
    "pytest-cov>=4.1.0",
    "black>=23.11.0",
    "isort>=5.12.0",
    "flake8>=6.1.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v"

[tool.black]
line-length = 88
target-version = ['py310']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
multi_line_output = 3
line_length = 88
known_first_party = ["training", "serving", "models", "scripts"]
